rename __main__ to _main
