+++ pyproject.toml
@@ -43,7 +43,7 @@ requires = ["setuptools>=60"]
 build-backend = "setuptools.build_meta"
 
 [project.scripts]
-jubilant = "jubilant.__main__:main"
+jubilant = "jubilant._main:main"
 
 [tool.setuptools.dynamic]
 version = {attr = "jubilant.__version__"}
